% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racism.R
\name{detect_hatespeech}
\alias{detect_hatespeech}
\title{Detect Dutch Racist Speech}
\usage{
detect_hatespeech(x, detector = racism_detector(), language = "dutch",
  type = c("tau", "udpipe"), anno = udpipe(x, language), detailed = FALSE)
}
\arguments{
\item{x}{a character vector of length 1 in UTF-8 encoding}

\item{detector}{an object of class \code{dutch_racism_terms} as returned by \code{racism_detector}}

\item{language}{the language. Defaults to 'dutch'}

\item{type}{character either udpipe or tau indicating to tokenise text with the udpipe R package or using the tau R package. Defaults to the tau package.}

\item{anno}{a data.frame with 1 row per term in \code{x} containing the columns token and lemma. Only used if type is set to udpipe.}

\item{detailed}{logical indicating to return a detailed description of each word in the character vector if it was racist + how many times it occurred (if type is tau).
Or to return the annotated udpipe dataset which indicates the 16 categories next to the whole text.}
}
\value{
a list with 16 elements indicating the counting of words in 16 racist categories
}
\description{
Detect Dutch Racist Speech in a character vector
}
\examples{
data(dutch_racistspeech)
racismtypes <- racism_detector(dutch_racistspeech, type = "original")

x <- "Eigen volk gaat voor, want die vuile manieren van de
EU moeten wij vanaf. Geen EU en geen VN. Waardeloos
en tegen onze mensen."
detect_hatespeech(x, detector = racismtypes, type = "tau")
detect_hatespeech(x, detector = racismtypes, type = "udpipe")

x <- "Burgemeester Termont is voor de zwartzakken die kiezen voor hem"
detect_hatespeech(x, detector = racismtypes, type = "tau")
detect_hatespeech(x, detector = racismtypes, type = "udpipe")

x <- "Oprotten die luizegaards"
detect_hatespeech(x, detector = racismtypes, type = "tau")
detect_hatespeech(x, detector = racismtypes, type = "udpipe")

x <- "Wil weer eens lukken dat wij met het vuilste krapuul
zitten, ik verschiet er zelfs niet van!"
detect_hatespeech(x, detector = racismtypes, type = "tau")
detect_hatespeech(x, detector = racismtypes, type = "udpipe")

x <- "Kan je niets aan doen dat je behoort tot het ras dat
nog minder verstand en gevoelens heeft in uw hersenen
dan het stinkend gat van een VARKEN ! :-p"
detect_hatespeech(x, detector = racismtypes, type = "tau")
detect_hatespeech(x, detector = racismtypes, type = "udpipe")

x <- "Het zijn precies de vreemden die de haat of het racisme
opwekken bij de autochtonen"
detect_hatespeech(x, detector = racismtypes, type = "tau")
detect_hatespeech(x, detector = racismtypes, type = "udpipe")
}
\seealso{
\code{\link{racism_detector}}
}
