% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racism.R
\name{racism_detector}
\alias{racism_detector}
\title{Create a racism detector}
\usage{
racism_detector(x, type = c("original", "cleaned", "expanded"))
}
\arguments{
\item{x}{a data.frame like the \code{dutch_racistspeech} inside this package}

\item{type}{either one of 'original', 'cleaned', 'expanded' to work on either one of the data}
}
\value{
an object of class \code{dutch_racism_terms} which is basically a list of words
which need to be found in the text as is and a regular expression containing hatespeech terms
}
\description{
Create a racism detector. This is basically a set of regular expressions
}
\examples{
racismtypes <- racism_detector(type = "expanded")

data(dutch_racistspeech)
racismtypes <- racism_detector(dutch_racistspeech, type = "original")
}
\seealso{
\code{\link{detect_racism}}, \code{\link{dutch_racistspeech}}
}
